separate ToolDetailsAbout into it's own source file
