+++ app/src/main/kotlin/org/cru/godtools/ui/tooldetails/ToolDetailsLayout.kt
@@ -1,6 +1,5 @@
 package org.cru.godtools.ui.tooldetails
 
-import android.text.util.Linkify
 import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement.spacedBy
@@ -13,12 +12,9 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Tab
 import androidx.compose.material3.TabRow
@@ -34,13 +30,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
@@ -54,24 +47,16 @@ import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
 import kotlinx.coroutines.launch
 import org.ccci.gto.android.common.androidx.compose.foundation.layout.padding
-import org.ccci.gto.android.common.androidx.compose.material3.ClickableText
 import org.ccci.gto.android.common.androidx.compose.material3.ui.tabs.pagerTabIndicatorOffset
-import org.ccci.gto.android.common.androidx.compose.material3.ui.text.addUriAnnotations
-import org.ccci.gto.android.common.androidx.compose.ui.text.getUriAnnotations
-import org.ccci.gto.android.common.compat.util.LocaleCompat
 import org.cru.godtools.R
 import org.cru.godtools.analytics.compose.RecordAnalyticsScreen
-import org.cru.godtools.analytics.model.ExitLinkActionEvent
-import org.cru.godtools.base.ui.compose.LocalEventBus
 import org.cru.godtools.base.ui.theme.GRAY_E6
 import org.cru.godtools.base.ui.theme.GT_RED
 import org.cru.godtools.base.ui.util.getFontFamilyOrNull
 import org.cru.godtools.base.ui.youtubeplayer.YouTubePlayer
 import org.cru.godtools.model.Tool
 import org.cru.godtools.model.Translation
-import org.cru.godtools.model.getDescription
 import org.cru.godtools.model.getName
-import org.cru.godtools.model.getSortedDisplayNames
 import org.cru.godtools.ui.tooldetails.analytics.model.ToolDetailsScreenEvent
 import org.cru.godtools.ui.tools.DownloadProgressIndicator
 import org.cru.godtools.ui.tools.PreloadTool
@@ -81,7 +66,6 @@ import org.cru.godtools.ui.tools.VariantToolCard
 private val TOOL_DETAILS_HORIZONTAL_MARGIN = 32.dp
 
 internal const val TEST_TAG_ACTION_TOOL_TRAINING = "action_tool_training"
-internal const val TEST_TAG_LANGUAGES_AVAILABLE = "languages_available"
 
 @Composable
 @OptIn(ExperimentalComposeUiApi::class, ExperimentalPagerApi::class)
@@ -276,33 +260,6 @@ internal fun ToolDetailsActions(
 }
 
 @Composable
-private fun ToolDetailsAbout(toolViewModel: ToolViewModels.ToolViewModel, modifier: Modifier = Modifier) {
-    val eventBus = LocalEventBus.current
-    val uriHandler = LocalUriHandler.current
-    val tool by toolViewModel.tool.collectAsState()
-    val translation by toolViewModel.firstTranslation.collectAsState()
-
-    ProvideTextStyle(MaterialTheme.typography.bodyMedium) {
-        Column(modifier = modifier) {
-            val description = remember { derivedStateOf { translation.value.getDescription(tool).orEmpty() } }.value
-                .addUriAnnotations(Linkify.WEB_URLS)
-            ClickableText(
-                description,
-                onClick = {
-                    description.getUriAnnotations(it, it).firstOrNull()?.let { (url) ->
-                        eventBus.post(ExitLinkActionEvent(toolViewModel.code, url, translation.value?.languageCode))
-                        uriHandler.openUri(url)
-                    }
-                },
-                fontFamily = translation.value?.getFontFamilyOrNull(),
-            )
-
-            ToolDetailsLanguages(toolViewModel, modifier = Modifier.padding(top = 48.dp))
-        }
-    }
-}
-
-@Composable
 private fun ToolDetailsVariants(
     viewModel: ToolDetailsFragmentDataModel,
     onVariantSelected: (String) -> Unit,
@@ -328,33 +285,5 @@ private fun ToolDetailsVariants(
                 onClick = { onVariantSelected(code) }
             )
         }
-    }
-}
-
-@Composable
-@VisibleForTesting
-internal fun ToolDetailsLanguages(viewModel: ToolViewModels.ToolViewModel, modifier: Modifier = Modifier) {
-    val languages by viewModel.availableLanguages.collectAsState()
-    if (languages.isEmpty()) return
-
-    val context = LocalContext.current
-    val locale = LocaleCompat.getDefault(LocaleCompat.Category.DISPLAY)
-    val displayLanguages by remember(context, locale) {
-        derivedStateOf { languages.getSortedDisplayNames(context, locale).joinToString(", ") }
-    }
-
-    Column(modifier = modifier.fillMaxWidth()) {
-        Text(
-            stringResource(R.string.tool_details_section_description_languages_available),
-            fontWeight = FontWeight.Bold
-        )
-        Divider(color = LocalContentColor.current.copy(alpha = 0.12f))
-        Text(
-            displayLanguages,
-            style = MaterialTheme.typography.bodyMedium,
-            modifier = Modifier
-                .testTag(TEST_TAG_LANGUAGES_AVAILABLE)
-                .padding(top = 8.dp)
-        )
     }
 }
